Review: Use enso-common
