+++ app/gui2/package.json
@@ -62,7 +62,6 @@
     "ag-grid-community": "^30.2.1",
     "ag-grid-enterprise": "^30.2.1",
     "codemirror": "^6.0.1",
-    "common-query": "file:../ide-desktop/lib/common-query",
     "culori": "^3.2.0",
     "enso-common": "file:../ide-desktop/lib/common",
     "enso-dashboard": "^0.1.0",
